requirements.txt
Django==1.11.6
django-appconf==1.0.2
django-imagekit==4.0.1
httplib2==0.10.3
numpy==1.13.3
oauth2==1.9.0.post1
olefile==0.44
opencv-python==3.3.0.10
pilkit==2.0
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.2
six==1.11.0
